% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{add_betting_cols}
\alias{add_betting_cols}
\title{Add Betting columns
This is only for DI-FBS football}
\usage{
add_betting_cols(play_df, g_id, yr)
}
\arguments{
\item{play_df}{(\emph{data.frame} required): Play-By-Play data.frame as pulled from `cfb_pbp_dat()` and `clean_pbp_dat()`)}

\item{g_id}{(\emph{Integer} optional): Game ID filter for querying a single game
Can be found using the `cfb_game_info()` function}

\item{yr}{(\emph{Integer} optional): Select year (example: 2018)}
}
\description{
Add Betting columns
This is only for DI-FBS football
}
\keyword{internal}
