% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{add_play_counts}
\alias{add_play_counts}
\title{Adds play counts to Play-by-Play data
Adds play counts to Play-by-Play data pulled from the API's raw game data}
\usage{
add_play_counts(play_df)
}
\arguments{
\item{play_df}{(\emph{data.frame} required): Adds play counts to Play-by-Play dataframe, as pulled from `cfb_pbp_dat()`}
}
\value{
The original `play_df` with the following columns appended/redefined:
\describe{
\item{game_play_number}{.}
\item{half_clock.minutes}{.}
\item{TimeSecsRem}{.}
\item{Under_two}{.}
\item{half}{.}
\item{kickoff_play}{.}
\item{receives_2H_kickoff}{.}
\item{score_diff}{.}
\item{lag_score_diff}{.}
\item{lag_offense_play}{.}
\item{lead_offense_play}{.}
\item{lead_offense_play2}{.}
\item{score_pts}{.}
\item{score_diff_start}{.}
\item{offense_receives_2H_kickoff}{.}
\item{half_play_number}{.}
\item{lag_off_timeouts}{.}
\item{lag_def_timeouts}{.}
\item{off_timeouts_rem_before}{.}
\item{def_timeouts_rem_before}{.}
\item{off_timeout_called}{.}
\item{def_timeout_called}{.}
\item{lead_TimeSecsRem}{.}
\item{end_of_half}{.}
\item{lead_play_type}{.}
\item{lead_play_type2}{.}
\item{lead_play_type3}{.}
\item{change_of_poss}{.}
}
}
\description{
Adds play counts to Play-by-Play data
Adds play counts to Play-by-Play data pulled from the API's raw game data
}
\details{
Requires the following columns to be present
\describe{
\item{game_id}{.}
\item{id_play}{.}
\item{clock.minutes}{.}
\item{clock.seconds}{.}
\item{half}{.}
\item{period}{.}
\item{offense_play}{.}
\item{defense_play}{.}
\item{home}{.}
\item{away}{.}
\item{offense_score}{.}
\item{defense_score}{.}
\item{offense_timeouts}{.}
\item{offense_timeouts}{.}
\item{play_text}{.}
\item{play_type}{.}
}
}
\keyword{internal}
