% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{add_timeout_cols}
\alias{add_timeout_cols}
\title{Add Timeout columns
This is only for DI-FBS football}
\usage{
add_timeout_cols(play_df)
}
\arguments{
\item{play_df}{(\emph{data.frame} required): Play-By-Play data.frame as pulled from `cfb_pbp_dat()` and `clean_pbp_dat()`)}
}
\description{
Add Timeout columns
This is only for DI-FBS football
}
\keyword{internal}
