% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{add_yardage}
\alias{add_yardage}
\title{Add yardage extracted from play text}
\usage{
add_yardage(play_df)
}
\arguments{
\item{play_df}{(\emph{data.frame} required) Extracts yardage information from Play-by-Play data frame, as pulled from `cfb_pbp_data()`}
}
\value{
The original `drive_df` with the following columns appended to it:
\describe{
\item{drive_id}{Returned as `drive_id` from original variable `drive_id`}
\item{pts_drive}{End result of the drive}
\item{scoring}{Logical flag for if drive was a scoring drive updated}
}
}
\description{
Add yardage extracted from play text
}
\details{
Cleans CFB (D-I) Drive-By-Drive Data to create `pts_drive` column. Requires the following columns be present:
\itemize{
\item{rush}{.}
\item{pass}{.}
\item{play_text}{Logical flag for if drive was a scoring drive}
\item{kickoff_play}{.}
\item{kickoff_play}{.}
\item{punt}{.}
\item{punt_tb}{.}
\item{punt_oob}{.}
\item{punt_downed}{.}
\item{punt_oob}{.}
\item{punt_oob}{.}
}
}
\keyword{internal}
