% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cfb_pbp_data.R
\name{add_yardage_select}
\alias{add_yardage_select}
\title{Select the columns needed for \code{add_yardage()}}
\usage{
add_yardage_select(pbp)
}
\arguments{
\item{pbp}{(\emph{data.frame} required) Play-by-Play dataframe pulled from API via the `cfb_pbp_data()` function}
}
\description{
Select the columns needed for \code{add_yardage()}
}
