% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_betting_lines.R
\name{cfb_betting_lines}
\alias{cfb_betting_lines}
\title{Get Betting information from games}
\source{
\url{https://api.collegefootballdata.com/lines}
}
\usage{
cfb_betting_lines(
  game_id = NULL,
  year = NULL,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  line_provider = NULL
)
}
\arguments{
\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game
Can be found using the \code{\link[=cfb_game_info]{cfb_game_info()}} function}

\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular or postseason}

\item{team}{(\emph{String} optional): D-I Team}

\item{home_team}{(\emph{String} optional): Home D-I Team}

\item{away_team}{(\emph{String} optional): Away D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{line_provider}{(\emph{String} optional): Select Line Provider - Caesars, consensus, numberfire, or teamrankings}
}
\value{
Betting information for games with the following columns:
\describe{
\item{\code{game_id}}{integer. Unique game identifier - \code{game_id}.}
\item{\code{season}}{integer. Season parameter.}
\item{\code{season_type}}{character. Season Type (regular, postseason, both).}
\item{\code{week}}{integer. Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier).}
\item{\code{home_team}}{character. Home D-I Team.}
\item{\code{home_conference}}{character. Home D-I Conference.}
\item{\code{home_score}}{integer. Home Score.}
\item{\code{away_team}}{character. Away D-I Team.}
\item{\code{away_conference}}{character. Away D-I Conference.}
\item{\code{away_score}}{integer. Away Score.}
\item{\code{provider}}{character. Line provider.}
\item{\code{spread}}{character. Spread for the game.}
\item{\code{formatted_spread}}{character. Formatted spread for the game.}
\item{\code{over_under}}{character. Over/Under for the game.}
}
}
\description{
Get Betting information from games
}
\examples{

cfb_betting_lines(year = 2018, week = 12, team = 'Florida State')

#7 OTs LSU at TAMU
cfb_betting_lines(year = 2018, week = 13, team = "Texas A&M", conference = 'SEC')

}
\keyword{Betting}
\keyword{Lines}
