% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_conf_types_df}
\alias{cfb_conf_types_df}
\title{College Football Conference}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{id}{Referencing conference id}
\item{name}{Conference name}
\item{short_name}{Short name for Conference}
\item{abbreviation}{Conference abbreviation}
...
}
}
\source{
\url{https://api.collegefootballdata.com}
}
\usage{
cfb_conf_types_df
}
\description{
This dataframes is used internally to identify conference names and use the abbreviation in the function call for data
}
\keyword{data}
