% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_conferences.R
\name{cfb_conferences}
\alias{cfb_conferences}
\title{CFB Conference Information}
\source{
\url{https://api.collegefootballdata.com/conferences}
}
\usage{
cfb_conferences()
}
\value{
A data frame with 11 rows and 4 variables:
\describe{
  \item{conference_id}{Referencing conference id}
  \item{name}{Conference name}
  \item{long_name}{Long name for Conference}
  \item{abbreviation}{Conference abbreviation}
  ...
}
}
\description{
Pulls all college football conferences and returns as data frame the following fields:
}
\examples{

cfb_conferences()

}
\keyword{Conferences}
