% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_espn_college_qbr.R
\name{cfb_espn_college_qbr_season}
\alias{cfb_espn_college_qbr_season}
\title{Get College Football QBR from ESPN API}
\usage{
cfb_espn_college_qbr_season(year = 2019)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}
}
\description{
Get College Football QBR from ESPN API
}
\examples{

cfb_espn_college_qbr_season(year=2018)

}
\keyword{College}
\keyword{ESPN}
\keyword{QBR}
