% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_fpi.R
\name{cfb_fpi_ranking}
\alias{cfb_fpi_ranking}
\title{Get FPI historical rating data (only end of season -- all that really matters for now)}
\usage{
cfb_fpi_ranking(season = 2019)
}
\arguments{
\item{season}{Year}
}
\description{
If conference is provided Get average S&P+ historical rating data by conference
}
\examples{


cfb_fpi_ranking(season=2018)
}
\keyword{internal}
