% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_box_advanced.R
\name{cfb_game_box_advanced}
\alias{cfb_game_box_advanced}
\title{Get game advanced box score information}
\usage{
cfb_game_box_advanced(game_id, long = FALSE)
}
\arguments{
\item{game_id}{(\emph{Integer} required): Game ID filter for querying a single game
Can be found using the \code{\link[cfbscrapR:cfb_game_info]{cfbscrapR::cfb_game_info()}} function}

\item{long}{(\emph{Logical} default `FALSE`): Return the data in a long format.}
}
\description{
Get game advanced box score information
}
\examples{

cfb_game_box_advanced(401114233)


}
\keyword{Advanced}
\keyword{Box}
\keyword{Game}
\keyword{Score}
