% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_info.R
\name{cfb_game_info}
\alias{cfb_game_info}
\title{Get results information from games}
\source{
\url{https://api.collegefootballdata.com/games}
}
\usage{
cfb_game_info(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  game_id = NULL,
  quarter_scores = FALSE
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{home_team}{(\emph{String} optional): Home D-I Team}

\item{away_team}{(\emph{String} optional): Away D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game\cr}

\item{quarter_scores}{(\emph{Logical} default FALSE): This is a parameter to return the
list columns that give the score at each quarter: home_line_scores and away_line scores.\cr
I have defaulted the parameter to false so that you will not have to go to the trouble of dropping it.}
}
\value{
A data frame with 22 variables:
\describe{
\item{\code{game_id}}{integer.}
\item{\code{season}}{integer.}
\item{\code{week}}{integer.}
\item{\code{season_type}}{character.}
\item{\code{start_date}}{character.}
\item{\code{start_time_tbd}}{logical.}
\item{\code{neutral_site}}{logical.}
\item{\code{conference_game}}{logical.}
\item{\code{attendance}}{integer.}
\item{\code{venue_id}}{integer.}
\item{\code{venue}}{character.}
\item{\code{home_id}}{integer.}
\item{\code{home_team}}{character.}
\item{\code{home_conference}}{character.}
\item{\code{home_points}}{integer.}
\item{\code{home_post_win_prob}}{character.}
\item{\code{away_id}}{integer.}
\item{\code{away_team}}{character.}
\item{\code{away_conference}}{character.}
\item{\code{away_points}}{integer.}
\item{\code{away_post_win_prob}}{character.}
\item{\code{excitement_index}}{character.}
}
}
\description{
Get results information from games
}
\examples{

cfb_game_info(2018, week = 1)

cfb_game_info(2018, week = 7, conference = 'Ind')

#7 OTs LSU @ TAMU
cfb_game_info(2018, week = 13, team = "Texas A&M", quarter_scores=TRUE)

}
\keyword{Game}
\keyword{Info}
