% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_media.R
\name{cfb_game_media}
\alias{cfb_game_media}
\title{Get Game media information (TV, radio, etc)}
\source{
\url{https://api.collegefootballdata.com/games/media}
}
\usage{
cfb_game_media(
  year,
  week = NULL,
  season_type = "both",
  team = NULL,
  conference = NULL,
  media_type = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default both): Select Season Type, regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{media_type}{(\emph{String} optional): Media type filter: tv, radio, web, ppv, or mobile}
}
\value{
A data frame with 13 variables:
\describe{
  \item{\code{game_id}}{integer.}
  \item{\code{season}}{integer.}
  \item{\code{week}}{integer.}
  \item{\code{season_type}}{character.}
  \item{\code{start_time}}{character.}
  \item{\code{is_start_time_tbd}}{logical.}
  \item{\code{home_team}}{character.}
  \item{\code{home_conference}}{character.}
  \item{\code{away_team}}{character.}
  \item{\code{away_conference}}{character.}
  \item{\code{tv}}{list.}
  \item{\code{radio}}{logical.}
  \item{\code{web}}{list.}
}
}
\description{
Get Game media information (TV, radio, etc)
}
\examples{

cfb_game_media(2019, week = 4, conference = 'ACC')

}
\keyword{Game}
\keyword{Info}
