% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_player_stats.R
\name{cfb_game_player_stats}
\alias{cfb_game_player_stats}
\title{Get results information from games}
\source{
\url{https://api.collegefootballdata.com/games/players}
}
\usage{
cfb_game_player_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  category = NULL,
  game_id = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular or postseason}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{category}{(\emph{String} optional): Category filter (e.g defensive)\cr
Offense: passing, receiving, rushing\cr
Defense: defensive, fumbles, interceptions\cr
Special Teams: punting, puntReturns, kicking, kickReturns\cr}

\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game
Can be found using the \code{\link[=cfb_game_info]{cfb_game_info()}} function}
}
\value{
A data frame with 32 variables:
\describe{
\item{\code{game_id}}{integer.}
\item{\code{team}}{character.}
\item{\code{conference}}{character.}
\item{\code{home_away}}{character.}
\item{\code{points}}{integer.}
\item{\code{category}}{character.}
\item{\code{athlete_id}}{character.}
\item{\code{name}}{character.}
\item{\code{c_att}}{character.}
\item{\code{yds}}{double.}
\item{\code{avg}}{double.}
\item{\code{td}}{double.}
\item{\code{int}}{double.}
\item{\code{qbr}}{double.}
\item{\code{car}}{double.}
\item{\code{long}}{double.}
\item{\code{rec}}{double.}
\item{\code{no}}{double.}
\item{\code{fg}}{character.}
\item{\code{pct}}{double.}
\item{\code{xp}}{character.}
\item{\code{pts}}{double.}
\item{\code{tb}}{double.}
\item{\code{in_20}}{double.}
\item{\code{fum}}{double.}
\item{\code{lost}}{double.}
\item{\code{tot}}{double.}
\item{\code{solo}}{double.}
\item{\code{sacks}}{double.}
\item{\code{tfl}}{double.}
\item{\code{pd}}{double.}
\item{\code{qb_hur}}{double.}
}
}
\description{
Get results information from games
}
\examples{

cfb_game_player_stats(2018, week = 15, conference = 'Ind')

cfb_game_player_stats(2013, week=1, team = "Florida State", category = 'passing')

}
\keyword{Game}
\keyword{Info}
