% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_records.R
\name{cfb_game_records}
\alias{cfb_game_records}
\title{Get Team records by year}
\usage{
cfb_game_records(year, team = NULL, conference = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{team}{(\emph{String} optional): Team - Select a valid team, D1 football}

\item{conference}{(\emph{String} optional): DI Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}
}
\description{
Get Team records by year
}
\examples{

cfb_game_records(2018, team = 'Notre Dame')


cfb_game_records(2013, team = "Florida State")

}
\keyword{Info}
\keyword{Team}
