% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_metrics_espn_wp.R
\name{cfb_metrics_espn_wp}
\alias{cfb_metrics_espn_wp}
\title{Get win probability chart data from ESPN}
\source{
\url{https://gist.github.com/mrcaseb/0f868193affb4be152e8e82c43a4dc07}
}
\usage{
cfb_metrics_espn_wp(game_id)
}
\arguments{
\item{game_id}{(\emph{Integer} required): Game ID filter for querying a single game\cr
Can be found using the \code{\link[cfbscrapR:cfb_game_info]{cfbscrapR::cfb_game_info()}} function}
}
\description{
Graciously contributed by MrCaseB:
}
\examples{

cfb_metrics_espn_wp(game_id = 401012356)


}
\keyword{Chart}
\keyword{Data}
\keyword{Probability}
\keyword{Win}
