% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_metrics_ppa_games.R
\name{cfb_metrics_ppa_games}
\alias{cfb_metrics_ppa_games}
\title{Get team game averages for Predicted Points Added (PPA)}
\usage{
cfb_metrics_ppa_games(
  year,
  week = NULL,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{excl_garbage_time}{(\emph{Logical} default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)}
}
\value{
A data frame with 18 variables:
\describe{
\item{\code{game_id}}{integer.}
\item{\code{season}}{integer.}
\item{\code{week}}{integer.}
\item{\code{conference}}{character.}
\item{\code{team}}{character.}
\item{\code{opponent}}{character.}
\item{\code{off_overall}}{character.}
\item{\code{off_passing}}{character.}
\item{\code{off_rushing}}{character.}
\item{\code{off_first_down}}{character.}
\item{\code{off_second_down}}{character.}
\item{\code{off_third_down}}{character.}
\item{\code{def_overall}}{character.}
\item{\code{def_passing}}{character.}
\item{\code{def_rushing}}{character.}
\item{\code{def_first_down}}{character.}
\item{\code{def_second_down}}{character.}
\item{\code{def_third_down}}{character.}
}
}
\description{
Get team game averages for Predicted Points Added (PPA)
}
\examples{

cfb_metrics_ppa_games(year = 2019, team = 'TCU')

}
\keyword{Points}
\keyword{Predicted}
\keyword{Teams}
