% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_metrics_wp.R
\name{cfb_metrics_wp}
\alias{cfb_metrics_wp}
\title{Get win probability chart data from API}
\source{
\url{https://api.collegefootballdata.com/metrics/wp}
}
\usage{
cfb_metrics_wp(game_id)
}
\arguments{
\item{game_id}{(\emph{Integer} required): Game ID filter for querying a single game\cr
Can be found using the \code{\link[=cfb_game_info]{cfb_game_info()}} function}
}
\value{
A data frame with 16 variables:
\describe{
\item{\code{play_id}}{character.}
\item{\code{play_text}}{character.}
\item{\code{home_id}}{integer.}
\item{\code{home}}{character.}
\item{\code{away_id}}{integer.}
\item{\code{away}}{character.}
\item{\code{spread}}{character.}
\item{\code{home_ball}}{logical.}
\item{\code{home_score}}{integer.}
\item{\code{away_score}}{integer.}
\item{\code{down}}{integer.}
\item{\code{distance}}{integer.}
\item{\code{home_win_prob}}{character.}
\item{\code{away_win_prob}}{double.}
\item{\code{play_number}}{integer.}
\item{\code{yard_line}}{integer.}
}
}
\description{
Get win probability chart data from API
}
\examples{

cfb_metrics_wp(game_id = 401012356)

}
\keyword{Chart}
\keyword{Data}
\keyword{Probability}
\keyword{Win}
