% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_play_stats_player.R
\name{cfb_play_stats_player}
\alias{cfb_play_stats_player}
\title{Gets player info associated by play}
\usage{
cfb_play_stats_player(
  year = NULL,
  week = NULL,
  team = NULL,
  game_id = NULL,
  athlete_id = NULL,
  stat_type_id = NULL,
  season_type = "regular"
)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{team}{(\emph{String} optional): D-I Team}

\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game\cr
Can be found using the \code{\link{cfb_game_info()}} function}

\item{athlete_id}{(\emph{Integer} optional): Athlete ID filter for querying a single athlete\cr
Can be found using the \code{\link{cfb_player_info()}} function.}

\item{stat_type_id}{(\emph{Integer} optional): Stat Type ID filter for querying a single stat type\cr
Can be found using the \code{\link{cfb_play_stats_types()}} function}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular, postseason, or both}
}
\description{
Information describes the players involved in the play
this includes passer, receiver, defensive players who
create sacks or picks, etc
}
\keyword{-}
\keyword{PBP}
\keyword{Player}
