% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_play_stats_types.R
\name{cfb_play_stats_types}
\alias{cfb_play_stats_types}
\title{College Football Mapping for Play Stats Types}
\source{
\url{https://api.collegefootballdata.com/play/stat/types}
}
\usage{
cfb_play_stats_types()
}
\value{
A data frame with 22 rows and 2 variables:
\describe{
\item{play_stat_type_id}{Referencing play stat type ID}
\item{name}{Type of play stats}
...
}
}
\description{
This function identifies all play stats types identified in the play by play data.
This can be used to filter out play stats types when calling functions before hand.
}
\examples{

cfb_play_stats_types()

}
\keyword{Plays}
