% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_play_type_df}
\alias{cfb_play_type_df}
\title{College Football Mapping for Play Types}
\format{
A data frame with 45 rows and 3 variables:
\describe{
\item{id}{Referencing play id}
\item{text}{play type description}
\item{abbreviation}{play type abbreviation used for function call}
...
}
}
\source{
\url{https://api.collegefootballdata.com}
}
\usage{
cfb_play_type_df
}
\description{
This data.frame helps identifies all play types identified in the play by play data.
This can be used to filter out play types when calling functions before hand.
}
\keyword{data}
