% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_player_returning.R
\name{cfb_player_returning}
\alias{cfb_player_returning}
\title{Player Information Search}
\usage{
cfb_player_returning(year = 2019, team = NULL, conference = NULL)
}
\arguments{
\item{year}{(\emph{Integer} required, default 2019): Year, 4 digit format (\emph{YYYY}).}

\item{team}{(\emph{String} optional): Team - Select a valid team, D1 football}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC}
}
\description{
A player usage function with \strong{Year} as a required input.
}
\examples{

cfb_player_returning(year = 2019, team = 'Florida State')


}
\keyword{Production}
\keyword{Returning}
