% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_player_usage.R
\name{cfb_player_usage}
\alias{cfb_player_usage}
\title{Player Information Search}
\usage{
cfb_player_usage(
  year = 2019,
  team = NULL,
  conference = NULL,
  position = NULL,
  athlete_id = NULL,
  excl_garbage_time = FALSE
)
}
\arguments{
\item{year}{(\emph{Integer} required, default 2019): Year, 4 digit format (\emph{YYYY}).}

\item{team}{(\emph{String} optional): Team - Select a valid team, D1 football}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{position}{(\emph{string} optional): Position of the player you are searching for.\cr
Position Group  - options include:\cr
 * Offense: QB, RB, FB, TE,  OL, G, OT, C, WR\cr
 * Defense: DB, CB, S, LB,  DE, DT, NT, DL\cr
 * Special Teams: K, P, LS, PK\cr}

\item{athlete_id}{(\emph{Integer} optional): Athlete ID filter for querying a single athlete
Can be found using the \code{\link[cfbscrapR:cfb_player_info]{cfbscrapR::cfb_player_info()}} function.}

\item{excl_garbage_time}{(\emph{Logical} default FALSE): Select whether to exclude Garbage Time (TRUE/FALSE)}
}
\description{
A player usage function with \strong{Year} as a required input.
}
\examples{

cfb_player_usage(year = 2019, position = 'WR', team = 'Florida State')


}
\keyword{Player}
\keyword{Usage}
