% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_rankings.R
\name{cfb_rankings}
\alias{cfb_rankings}
\title{Gets Historical CFB poll rankings at a specific week}
\usage{
cfb_rankings(year, week = NULL, season_type = "regular")
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default regular): Season type - regular or postseason}
}
\description{
Postseason polls are after Week 13
}
\examples{

cfb_rankings(year = 2019, week=12)

cfb_rankings(year = 2018, week = 14)

cfb_rankings(year = 2013, season_type = 'postseason')

}
\keyword{CFB}
\keyword{Rankings}
