% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_fpi.R
\name{cfb_ratings_fpi}
\alias{cfb_ratings_fpi}
\title{Get FPI historical rating data (most recent of each season)}
\usage{
cfb_ratings_fpi(season = 2019)
}
\arguments{
\item{season}{Year}
}
\description{
Get FPI historical rating data (most recent of each season)
}
\examples{


cfb_ratings_fpi(season=2018)
}
\keyword{internal}
