% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_sp_conference.R
\name{cfb_ratings_sp_conference}
\alias{cfb_ratings_sp_conference}
\title{Get conference-level S&P+ historical rating data}
\source{
\url{https://api.collegefootballdata.com/ratings/sp/conferences}
}
\usage{
cfb_ratings_sp_conference(year = NULL, conference = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{conference}{(\emph{String} optional): Conference abbreviation - S&P+ information by conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}
}
\value{
A data frame with 25 variables:
\describe{
  \item{\code{year}}{integer.}
  \item{\code{conference}}{character.}
  \item{\code{rating}}{double.}
  \item{\code{second_order_wins}}{logical.}
  \item{\code{sos}}{logical.}
  \item{\code{offense_rating}}{double.}
  \item{\code{offense_success}}{logical.}
  \item{\code{offense_explosiveness}}{logical.}
  \item{\code{offense_rushing}}{logical.}
  \item{\code{offense_passing}}{logical.}
  \item{\code{offense_standard_downs}}{logical.}
  \item{\code{offense_passing_downs}}{logical.}
  \item{\code{offense_run_rate}}{logical.}
  \item{\code{offense_pace}}{logical.}
  \item{\code{defense_rating}}{double.}
  \item{\code{defense_success}}{logical.}
  \item{\code{defense_explosiveness}}{logical.}
  \item{\code{defense_rushing}}{logical.}
  \item{\code{defense_passing}}{logical.}
  \item{\code{defense_standard_downs}}{logical.}
  \item{\code{defense_passing_downs}}{logical.}
  \item{\code{defense_havoc_total}}{logical.}
  \item{\code{defense_havoc_front_seven}}{logical.}
  \item{\code{defense_havoc_db}}{logical.}
  \item{\code{special_teams_rating}}{double.}
}
}
\description{
Get conference-level S&P+ historical rating data
}
\examples{

cfb_ratings_sp_conference(year = 2019)

cfb_ratings_sp_conference(year = 2012, conference = 'SEC')

cfb_ratings_sp_conference(year = 2016, conference = 'ACC')

}
\keyword{SP+}
