% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_srs.R
\name{cfb_ratings_srs}
\alias{cfb_ratings_srs}
\title{Get SRS historical rating data}
\usage{
cfb_ratings_srs(year = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference name - select a valid FBS conference\cr
Conference names P5: ACC,  Big 12, Big Ten, SEC, Pac-12\cr
Conference names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic\cr}
}
\description{
At least one of \strong{year} or \strong{team} must be specified for the function to run
}
\examples{

cfb_ratings_srs(year = 2019, team = "Texas")

cfb_ratings_srs(year = 2018, conference = 'SEC')

}
\keyword{SRS}
