% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_recruiting_player.R
\name{cfb_recruiting_player}
\alias{cfb_recruiting_player}
\title{CFB Recruiting
Gets CFB recruiting information for a single year with filters available for team,
recruit type, state and position.}
\usage{
cfb_recruiting_player(
  year = NULL,
  team = NULL,
  recruit_type = "HighSchool",
  state = NULL,
  position = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY}) - Minimum: 2000, Maximum: 2020 currently}

\item{team}{(\emph{String} optional): D-I Team}

\item{recruit_type}{(\emph{String} optional): default API return is 'HighSchool', other options include 'JUCO'
or 'PrepSchool'  - For position group information}

\item{state}{(\emph{String} optional): Two letter State abbreviation}

\item{position}{(\emph{String} optional): Position Group  - options include:\cr
* Offense: 'PRO', 'DUAL', 'RB', 'FB', 'TE',  'OT', 'OG', 'OC', 'WR'\cr
* Defense: 'CB', 'S', 'OLB', 'ILB', 'WDE', 'SDE', 'DT'\cr
* Special Teams: 'K', 'P'\cr}
}
\description{
At least one of \strong{year} or \strong{team} must be specified for the function to run
}
\details{
If you would like CFB recruiting information for teams, please see the \code{\link[cfbscrapR:cfb_recruiting_team]{cfbscrapR::cfb_recruiting_team()}}  function

If you would like to get cfb recruiting information based on position groups during a
time period for all FBS teams, please see the \code{\link[cfbscrapR:cfb_recruiting_position]{cfbscrapR::cfb_recruiting_position()}} function.
}
\examples{

cfb_recruiting_player(2018, team = "Texas")

cfb_recruiting_player(2016, recruit_type = 'JUCO')

cfb_recruiting_player(2020, recruit_type = 'HighSchool', position = 'OT', state = 'FL')

}
\keyword{Recruiting}
