% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_recruiting_position.R
\name{cfb_recruiting_position}
\alias{cfb_recruiting_position}
\title{CFB Recruiting Information - Position Groups}
\usage{
cfb_recruiting_position(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL
)
}
\arguments{
\item{start_year}{(\emph{Integer} optional): Start Year, 4 digit format (\emph{YYYY}). \emph{Note: 2000 is the minimum value}}

\item{end_year}{(\emph{Integer} optional): End Year,  4 digit format (\emph{YYYY}). \emph{Note: 2020 is the maximum value currently}}

\item{team}{(\emph{String} optional): Team - Select a valid team, D-I football}

\item{conference}{(\emph{String} optional): Conference name - select a valid FBS conference
Conference names P5: ACC,  Big 12, Big Ten, SEC, Pac-12
Conference names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic}
}
\description{
If only start_year is provided, function will get CFB recruiting information based
on position groups during that year for all FBS teams.
}
\examples{

cfb_recruiting_position(2018, team="Texas")

cfb_recruiting_position(2016, 2020, team="Virginia")

cfb_recruiting_position(2015, 2020, conference = "SEC")

}
\keyword{Recruiting}
