% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_recruiting_team.R
\name{cfb_recruiting_team}
\alias{cfb_recruiting_team}
\title{CFB Recruiting Information - Team Rankings}
\usage{
cfb_recruiting_team(year = NULL, team = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Recruiting Class Year, 4 digit format (\emph{YYYY}). \emph{Note: 2000 is the minimum value}}

\item{team}{(\emph{String} optional): Team - Select a valid team, D1 football}
}
\description{
Gets CFB team recruiting ranks with filters available for year and team.
At least one of \strong{year} or \strong{team} must be specified for the function to run
}
\details{
If you would like CFB recruiting information for players, please
see the \code{\link[cfbscrapR:cfb_recruiting_player]{cfbscrapR::cfb_recruiting_player()}} function

If you would like to get CFB recruiting information based on position groups during a
time period for all FBS teams, please see the \code{\link[cfbscrapR:cfb_recruiting_position]{cfbscrapR::cfb_recruiting_position()}} function.
}
\examples{

cfb_recruiting_team(2018, team = "Texas")

cfb_recruiting_team(2016, team = "Virginia")

cfb_recruiting_team(2016, team = "Texas A&M")

cfb_recruiting_team(2011)

}
\keyword{Recruiting}
