% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_categories.R
\name{cfb_stats_categories}
\alias{cfb_stats_categories}
\title{College Football Mapping for Stats Categories}
\source{
\url{https://api.collegefootballdata.com/stats/categories}
}
\usage{
cfb_stats_categories()
}
\value{
A data frame with 38 values:
\describe{
\item{name}{Statistics Categories}
...
}
}
\description{
This function identifies all Stats Categories identified in the regular stats endpoint.
}
\examples{

cfb_stats_categories()

}
\keyword{Categories}
\keyword{Stats}
