% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_game_advanced.R
\name{cfb_stats_game_advanced}
\alias{cfb_stats_game_advanced}
\title{Get Game Advanced Stats}
\source{
\url{https://api.collegefootballdata.com/stats/game/advanced}
}
\usage{
cfb_stats_game_advanced(
  year,
  week = NULL,
  team = NULL,
  opponent = NULL,
  excl_garbage_time = FALSE,
  season_type = "both"
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{team}{(\emph{String} optional): D-I Team}

\item{opponent}{(\emph{String} optional): Opponent D-I Team}

\item{excl_garbage_time}{(\emph{Logical} default FALSE): Select whether to exclude Garbage Time (TRUE/FALSE)}

\item{season_type}{(\emph{String} default both): Select Season Type: regular, postseason, or both.}
}
\value{
A data frame with 60 variables:
\describe{
  \item{\code{game_id}}{integer.}
  \item{\code{week}}{integer.}
  \item{\code{team}}{character.}
  \item{\code{opponent}}{character.}
  \item{\code{off_plays}}{integer.}
  \item{\code{off_drives}}{integer.}
  \item{\code{off_ppa}}{double.}
  \item{\code{off_total_ppa}}{double.}
  \item{\code{off_success_rate}}{double.}
  \item{\code{off_explosiveness}}{double.}
  \item{\code{off_power_success}}{double.}
  \item{\code{off_stuff_rate}}{double.}
  \item{\code{off_line_yds}}{double.}
  \item{\code{off_line_yds_total}}{integer.}
  \item{\code{off_second_lvl_yds}}{double.}
  \item{\code{off_second_lvl_yds_total}}{integer.}
  \item{\code{off_open_field_yds}}{integer.}
  \item{\code{off_open_field_yds_total}}{integer.}
  \item{\code{off_standard_downs_ppa}}{double.}
  \item{\code{off_standard_downs_success_rate}}{double.}
  \item{\code{off_standard_downs_explosiveness}}{double.}
  \item{\code{off_passing_downs_ppa}}{double.}
  \item{\code{off_passing_downs_success_rate}}{double.}
  \item{\code{off_passing_downs_explosiveness}}{double.}
  \item{\code{off_rushing_plays_ppa}}{double.}
  \item{\code{off_rushing_plays_total_ppa}}{double.}
  \item{\code{off_rushing_plays_success_rate}}{double.}
  \item{\code{off_rushing_plays_explosiveness}}{double.}
  \item{\code{off_passing_plays_ppa}}{double.}
  \item{\code{off_passing_plays_total_ppa}}{double.}
  \item{\code{off_passing_plays_success_rate}}{double.}
  \item{\code{off_passing_plays_explosiveness}}{double.}
  \item{\code{def_plays}}{integer.}
  \item{\code{def_drives}}{integer.}
  \item{\code{def_ppa}}{double.}
  \item{\code{def_total_ppa}}{double.}
  \item{\code{def_success_rate}}{double.}
  \item{\code{def_explosiveness}}{double.}
  \item{\code{def_power_success}}{double.}
  \item{\code{def_stuff_rate}}{double.}
  \item{\code{def_line_yds}}{double.}
  \item{\code{def_line_yds_total}}{integer.}
  \item{\code{def_second_lvl_yds}}{double.}
  \item{\code{def_second_lvl_yds_total}}{integer.}
  \item{\code{def_open_field_yds}}{double.}
  \item{\code{def_open_field_yds_total}}{integer.}
  \item{\code{def_standard_downs_ppa}}{double.}
  \item{\code{def_standard_downs_success_rate}}{double.}
  \item{\code{def_standard_downs_explosiveness}}{double.}
  \item{\code{def_passing_downs_ppa}}{double.}
  \item{\code{def_passing_downs_success_rate}}{double.}
  \item{\code{def_passing_downs_explosiveness}}{double.}
  \item{\code{def_rushing_plays_ppa}}{double.}
  \item{\code{def_rushing_plays_total_ppa}}{double.}
  \item{\code{def_rushing_plays_success_rate}}{double.}
  \item{\code{def_rushing_plays_explosiveness}}{double.}
  \item{\code{def_passing_plays_ppa}}{double.}
  \item{\code{def_passing_plays_total_ppa}}{double.}
  \item{\code{def_passing_plays_success_rate}}{double.}
  \item{\code{def_passing_plays_explosiveness}}{double.}
}
}
\description{
Get Game Advanced Stats
}
\examples{

cfb_stats_game_advanced(year = 2018, week = 12, team = 'Texas A&M')

cfb_stats_game_advanced(2019, team = 'LSU')

cfb_stats_game_advanced(2013, team = "Florida State")

}
\keyword{Advanced}
\keyword{Game}
\keyword{Stats}
