% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_season_advanced.R
\name{cfb_stats_season_advanced}
\alias{cfb_stats_season_advanced}
\title{Get Season Advanced Statistics by Team}
\usage{
cfb_stats_season_advanced(
  year,
  team = NULL,
  excl_garbage_time = FALSE,
  start_week = NULL,
  end_week = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{team}{(\emph{String} optional): D-I Team}

\item{excl_garbage_time}{(\emph{Logical} default FALSE): Select whether to exclude Garbage Time (TRUE/FALSE)}

\item{start_week}{(\emph{Integer} optional): Starting Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{end_week}{(\emph{Integer} optional): Ending Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}
}
\description{
Get Season Advanced Statistics by Team
}
\examples{

cfb_stats_season_advanced(2019, team = 'LSU')


}
\keyword{Advanced}
\keyword{Season}
\keyword{Stats}
\keyword{Team}
