% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_season_player.R
\name{cfb_stats_season_player}
\alias{cfb_stats_season_player}
\title{Get Season Statistics by Player}
\source{
\url{https://api.collegefootballdata.com/stats/player/season}
}
\usage{
cfb_stats_season_player(
  year,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  start_week = NULL,
  end_week = NULL,
  category = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{season_type}{(\emph{String} default: regular): Select Season Type - regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{start_week}{(\emph{Integer} optional): Starting Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{end_week}{(\emph{Integer} optional): Ending Week - values range fom 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{category}{(\emph{String} optional): Category filter (e.g defensive)\cr
Offense: passing, receiving, rushing\cr
Defense: defensive, fumbles, interceptions\cr
Special Teams: punting, puntReturns, kicking, kickReturns\cr}
}
\value{
A data frame with 59 variables:
\describe{
  \item{\code{team}}{character.}
  \item{\code{conference}}{character.}
  \item{\code{athlete_id}}{character.}
  \item{\code{player}}{character.}
  \item{\code{category}}{character.}
  \item{\code{passing_completions}}{double.}
  \item{\code{passing_att}}{double.}
  \item{\code{passing_pct}}{double.}
  \item{\code{passing_yds}}{double.}
  \item{\code{passing_td}}{double.}
  \item{\code{passing_int}}{double.}
  \item{\code{passing_ypa}}{double.}
  \item{\code{rushing_car}}{double.}
  \item{\code{rushing_yds}}{double.}
  \item{\code{rushing_td}}{double.}
  \item{\code{rushing_ypc}}{double.}
  \item{\code{rushing_long}}{double.}
  \item{\code{receiving_rec}}{double.}
  \item{\code{receiving_yds}}{double.}
  \item{\code{receiving_td}}{double.}
  \item{\code{receiving_ypr}}{double.}
  \item{\code{receiving_long}}{double.}
  \item{\code{fumbles_fum}}{double.}
  \item{\code{fumbles_rec}}{double.}
  \item{\code{fumbles_lost}}{double.}
  \item{\code{defensive_solo}}{double.}
  \item{\code{defensive_tot}}{double.}
  \item{\code{defensive_tfl}}{double.}
  \item{\code{defensive_sacks}}{double.}
  \item{\code{defensive_qb_hur}}{double.}
  \item{\code{interceptions_int}}{double.}
  \item{\code{interceptions_yds}}{double.}
  \item{\code{interceptions_avg}}{double.}
  \item{\code{interceptions_td}}{double.}
  \item{\code{defensive_pd}}{double.}
  \item{\code{defensive_td}}{double.}
  \item{\code{kicking_fgm}}{double.}
  \item{\code{kicking_fga}}{double.}
  \item{\code{kicking_pct}}{double.}
  \item{\code{kicking_xpa}}{double.}
  \item{\code{kicking_xpm}}{double.}
  \item{\code{kicking_pts}}{double.}
  \item{\code{kicking_long}}{double.}
  \item{\code{kick_returns_no}}{double.}
  \item{\code{kick_returns_yds}}{double.}
  \item{\code{kick_returns_avg}}{double.}
  \item{\code{kick_returns_td}}{double.}
  \item{\code{kick_returns_long}}{double.}
  \item{\code{punting_no}}{double.}
  \item{\code{punting_yds}}{double.}
  \item{\code{punting_ypp}}{double.}
  \item{\code{punting_long}}{double.}
  \item{\code{punting_in_20}}{double.}
  \item{\code{punting_tb}}{double.}
  \item{\code{punt_returns_no}}{double.}
  \item{\code{punt_returns_yds}}{double.}
  \item{\code{punt_returns_avg}}{double.}
  \item{\code{punt_returns_td}}{double.}
  \item{\code{punt_returns_long}}{double.}
}
}
\description{
Get Season Statistics by Player
}
\examples{

cfb_stats_season_player(year=2018, conference = 'B12', start_week = 1, end_week = 7)

cfb_stats_season_player(2019, team = 'LSU', category = 'passing')

cfb_stats_season_player(2013, team = "Florida State", category = 'passing')

}
\keyword{Player}
\keyword{Season}
\keyword{Stats}
