% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_season_team.R
\name{cfb_stats_season_team}
\alias{cfb_stats_season_team}
\title{Get Season Statistics by Team}
\source{
\url{https://api.collegefootballdata.com/stats/season}
}
\usage{
cfb_stats_season_team(
  year,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  start_week = NULL,
  end_week = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{season_type}{(\emph{String} default: regular): Select Season Type - regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{start_week}{(\emph{Integer} optional): Starting Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{end_week}{(\emph{Integer} optional): Ending Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}
}
\value{
A data frame with 46 variables:
\describe{
\item{\code{games}}{integer.}
\item{\code{team}}{character.}
\item{\code{conference}}{character.}
\item{\code{games}}{integer.}
\item{\code{time_of_poss_total}}{integer.}
\item{\code{time_of_poss_pg}}{double.}
\item{\code{pass_comps}}{integer.}
\item{\code{pass_atts}}{integer.}
\item{\code{completion_pct}}{double.}
\item{\code{net_pass_yds}}{integer.}
\item{\code{pass_ypa}}{double.}
\item{\code{pass_ypr}}{double.}
\item{\code{pass_TDs}}{integer.}
\item{\code{interceptions}}{integer.}
\item{\code{int_pct}}{double.}
\item{\code{rush_atts}}{integer.}
\item{\code{rush_yds}}{integer.}
\item{\code{rush_TDs}}{integer.}
\item{\code{rush_ypc}}{double.}
\item{\code{total_yds}}{integer.}
\item{\code{fumbles_lost}}{integer.}
\item{\code{turnovers}}{integer.}
\item{\code{turnovers_pg}}{double.}
\item{\code{first_downs}}{integer.}
\item{\code{third_downs}}{integer.}
\item{\code{third_down_convs}}{integer.}
\item{\code{third_conv_rate}}{double.}
\item{\code{fourth_down_convs}}{integer.}
\item{\code{fourth_downs}}{integer.}
\item{\code{fourth_conv_rate}}{double.}
\item{\code{penalties}}{integer.}
\item{\code{penalty_yds}}{integer.}
\item{\code{penalties_pg}}{double.}
\item{\code{penalty_yds_pg}}{double.}
\item{\code{yards_per_penalty}}{double.}
\item{\code{kick_returns}}{integer.}
\item{\code{kick_return_yds}}{integer.}
\item{\code{kick_return_TDs}}{integer.}
\item{\code{kick_return_avg}}{double.}
\item{\code{punt_returns}}{integer.}
\item{\code{punt_return_yds}}{integer.}
\item{\code{punt_return_TDs}}{integer.}
\item{\code{punt_return_avg}}{double.}
\item{\code{passes_intercepted}}{integer.}
\item{\code{passes_intercepted_yds}}{integer.}
\item{\code{passes_intercepted_TDs}}{integer.}
}
}
\description{
Get Season Statistics by Team
}
\examples{

cfb_stats_season_team(year=2018, conference = 'B12', start_week = 1,end_week = 8)

cfb_stats_season_team(2019, team = 'LSU')

cfb_stats_season_team(2013, team = "Florida State")

}
\keyword{Season}
\keyword{Stats}
\keyword{Team}
