% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_team_roster.R
\name{cfb_team_roster}
\alias{cfb_team_roster}
\title{Team Roster
Get a teams full roster by year. If year not selected, API defaults to most recent year (2019 as of 6/23/20)}
\usage{
cfb_team_roster(team, year = NULL)
}
\arguments{
\item{team}{(\emph{String} required): Team, select a valid team in D-I football}

\item{year}{(\emph{Integer} optional): Year,  4 digit format (\emph{YYYY})}
}
\description{
Team Roster
Get a teams full roster by year. If year not selected, API defaults to most recent year (2019 as of 6/23/20)
}
\examples{

cfb_team_roster("Florida State")

}
\keyword{Roster}
\keyword{Team}
