% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_venues.R
\name{cfb_venues}
\alias{cfb_venues}
\title{CFB Venue Information}
\usage{
cfb_venues()
}
\description{
Pulls all college football venues and data on capacity, grass, city/state, location,
elevation, dome, timezone and construction year
}
\examples{

cfb_venues()

}
\keyword{Venues}
