% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{clean_drive_info}
\alias{clean_drive_info}
\title{Clean Drive Information
Cleans CFB (D-I) Drive-By-Drive Data to create `pts_drive` column}
\usage{
clean_drive_info(drive_df)
}
\arguments{
\item{drive_df}{(\emph{data.frame} required) Drive DataFrame pulled from API}
}
\description{
Clean Drive Information
Cleans CFB (D-I) Drive-By-Drive Data to create `pts_drive` column
}
\keyword{internal}
