% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{clean_pbp_dat}
\alias{clean_pbp_dat}
\title{Clean Play-by-Play data
Cleans Play-by-Play data pulled from the API's raw game data}
\usage{
clean_pbp_dat(raw_df)
}
\arguments{
\item{raw_df}{(\emph{data.frame} required): Performs data cleansing on Play-by-Play DataFrame, as pulled from `cfb_pbp_dat()`}
}
\description{
Clean Play-by-Play data
Cleans Play-by-Play data pulled from the API's raw game data
}
\keyword{internal}
