% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cfb_pbp_data.R
\name{clean_pbp_dat_select}
\alias{clean_pbp_dat_select}
\title{Select the columns needed for \code{clean_pbp_dat()}}
\usage{
clean_pbp_dat_select(pbp)
}
\arguments{
\item{pbp}{(\emph{data.frame} required) Play-by-Play dataframe pulled from API via the `cfb_pbp_data()` function}
}
\description{
Select the columns needed for \code{clean_pbp_dat()}
}
