% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wpa_betting.R
\name{create_wpa_betting}
\alias{create_wpa_betting}
\title{Add Win Probability Added (WPA) calculations to Play-by-Play DataFrame (Betting)
This is only for D1 football}
\usage{
create_wpa_betting(df, wp_model = cfbscrapR:::wp_model)
}
\arguments{
\item{df}{(\emph{data.frame} required): Clean Play-by-Play data.frame with Expected Points Added (EPA) calculations}

\item{wp_model}{(\emph{model} default `cfbscrapR:::wp_model`): Win Probability (WP) Model}
}
\description{
Extracts raw game by game data.
}
\keyword{internal}
