% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_calc_ep_wp.R
\name{ep_model_select_check}
\alias{ep_model_select_check}
\title{Select the columns needed for EP predictions}
\usage{
ep_model_select_check(pbp)
}
\arguments{
\item{pbp}{(\emph{data.frame} required) Play-by-Play dataframe pulled from API via the `cfb_pbp_data()` function}
}
\description{
Select the columns needed for EP predictions
}
