% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_calc_ep_wp.R
\name{get_preds_ep}
\alias{get_preds_ep}
\title{Select the columns needed for EP predictions}
\usage{
get_preds_ep(pbp, ep_model = cfbscrapR:::ep_model)
}
\arguments{
\item{pbp}{(\emph{data.frame} required) Play-by-Play dataframe pulled from API via the `cfb_pbp_data()` function}

\item{ep_model}{(\emph{model} default cfbscrapR:::ep_model): Expected Points (EP) Model}
}
\description{
Select the columns needed for EP predictions
}
