% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{penalty_detection}
\alias{penalty_detection}
\title{Penalty Detection
Adds penalty columns to Play-by-Play data pulled from the API}
\usage{
penalty_detection(raw_df)
}
\arguments{
\item{raw_df}{(\emph{data.frame} required): Performs data cleansing on Play-by-Play DataFrame, as pulled from `cfb_pbp_dat()`}
}
\description{
Penalty Detection
Adds penalty columns to Play-by-Play data pulled from the API
}
\keyword{internal}
