% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_away_travel.R
\name{plot_away_travel}
\alias{plot_away_travel}
\title{Plot a teams away travel}
\usage{
plot_away_travel(year, team_name)
}
\arguments{
\item{year}{Year}

\item{team_name}{Team Name}
}
\description{
Plot a teams away travel
}
\examples{

plot_away_travel(2019,"Texas")



}
\keyword{Plot}
\keyword{Travel}
