% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pbp_sequencing.R
\name{plot_pbp_sequencing}
\alias{plot_pbp_sequencing}
\title{Plot the Drive/PBP Sequencing}
\usage{
plot_pbp_sequencing(df)
}
\arguments{
\item{df}{Play-by-Play data frame (can be retrieved from \code{\link[=cfb_pbp_data]{cfb_pbp_data()}})}
}
\description{
Plot the Drive/PBP Sequencing
}
\examples{

df <- cfb_pbp_data(year = 2019, week = 9,team = 'Texas A&M', epa_wpa = TRUE)

plot_pbp_sequencing(df)

}
\keyword{PbP}
\keyword{Plot}
