% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wpa.R
\name{plot_wpa}
\alias{plot_wpa}
\title{Plot the WPA for a specific game}
\usage{
plot_wpa(
  dat,
  game_id = NULL,
  away_color,
  home_color,
  winner = "home",
  bet = TRUE
)
}
\arguments{
\item{dat}{(\emph{data.frame} required) Play-by-Play data.frame as can be retrieved from \code{\link[cfbscrapR:cfb_pbp_data]{cfbscrapR::cfb_pbp_data()}}}

\item{game_id}{(\emph{Integer} optional) Game ID filter for querying a single game
Can be found using the \code{\link[cfbscrapR:cfb_game_info]{cfbscrapR::cfb_game_info()}} function}

\item{away_color}{color selection for the away team}

\item{home_color}{color selection for the home team}

\item{winner}{Winner of contest, home or away.}

\item{bet}{TRUE or FALSE to use initial Win Probability converted from betting lines.}
}
\description{
Plot the WPA for a specific game
}
\examples{
df = cfb_pbp_data(year = 2019, week = 12, team = 'Baylor', epa_wpa = TRUE)
plot_wpa(df, away_color=c(OU="#841617"), home_color=c(BU="#003015"), winner="away")

}
\keyword{Plot}
\keyword{WPA}
