% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pbp_sequencing.R
\name{prep_df_pbp_overview}
\alias{prep_df_pbp_overview}
\title{prep_df_pbp_overview}
\usage{
prep_df_pbp_overview(df)
}
\arguments{
\item{df}{Play-by-Play data frame (can be retrieved from cfb_pbp_data)}
}
\description{
prep_df_pbp_overview
}
\keyword{PBP}
\keyword{Plot}
