% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cfb_pbp_data.R
\name{prep_epa_df_after_select}
\alias{prep_epa_df_after_select}
\title{Select the columns needed for \code{prep_epa_df_after()}}
\usage{
prep_epa_df_after_select(pbp)
}
\arguments{
\item{pbp}{(\emph{data.frame} required) Play-by-Play dataframe pulled from API via the `cfb_pbp_data()` function}
}
\description{
Select the columns needed for \code{prep_epa_df_after()}
}
