% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epa.R
\name{prep_epa_df_before}
\alias{prep_epa_df_before}
\title{Prep for EPA calculations at the start of the play
This is only for D1 football}
\usage{
prep_epa_df_before(df)
}
\arguments{
\item{df}{(\emph{data.frame} required) Clean Play-by-Play DataFrame pulled from `cfb_pbp_dat()`}
}
\description{
Extracts raw game by game data.
}
\keyword{internal}
