% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_calc_ep_wp.R
\name{wp_model_select_check}
\alias{wp_model_select_check}
\title{Select the columns needed for WP predictions}
\usage{
wp_model_select_check(pbp)
}
\arguments{
\item{pbp}{(\emph{data.frame} required) Play-by-Play dataframe pulled from API via the `cfb_pbp_data()` function}
}
\description{
Select the columns needed for WP predictions
}
