% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wpa.R
\name{wpa_calcs}
\alias{wpa_calcs}
\title{WPA Calcs}
\usage{
wpa_calcs(df)
}
\arguments{
\item{df}{(\emph{data.frame} required): Clean Play-by-Play data.frame with Expected Points Added (EPA) calculations}
}
\description{
Extracts raw game by game data.
}
\keyword{internal}
