% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_betting_lines.R
\name{cfb_betting_lines}
\alias{cfb_betting_lines}
\title{Get Betting information from games}
\usage{
cfb_betting_lines(
  game_id = NULL,
  year = NULL,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  line_provider = NULL
)
}
\arguments{
\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game
Can be found using the `cfb_game_info()` function}

\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular or postseason}

\item{team}{(\emph{String} optional): D-I Team}

\item{home_team}{(\emph{String} optional): Home D-I Team}

\item{away_team}{(\emph{String} optional): Away D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{line_provider}{(\emph{String} optional): Select Line Provider - Caesars, consensus, numberfire, or teamrankings}
}
\description{
Get Betting information from games
}
\examples{

cfb_betting_lines(year = 2018, week = 12, team = 'Florida State')

#7 OTs LSU at TAMU
cfb_betting_lines(year = 2018, week = 13, team = "Texas A&M", conference = 'SEC')


}
\keyword{Betting}
\keyword{Lines}
