% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_drives.R
\name{cfb_drives}
\alias{cfb_drives}
\title{Get results information from games}
\usage{
cfb_drives(
  year,
  season_type = "regular",
  week = NULL,
  team = NULL,
  offense_team = NULL,
  defense_team = NULL,
  conference = NULL,
  offense_conference = NULL,
  defense_conference = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular, postseason, or both}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{team}{(\emph{String} optional): D-I Team}

\item{offense_team}{(\emph{String} optional): Offense D-I Team}

\item{defense_team}{(\emph{String} optional): Defense D-I Team}

\item{conference}{(\emph{String} optional): DI Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{offense_conference}{(\emph{String} optional): Offense DI Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{defense_conference}{(\emph{String} optional): Defense DI Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}
}
\description{
Get results information from games
}
\examples{

cfb_drives(2018, week = 1, team = "TCU")

cfb_drives(2018, team = "Texas A&M", defense_conference = 'SEC')

}
\keyword{Drives}
