% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_box_advanced.R
\name{cfb_game_box_advanced}
\alias{cfb_game_box_advanced}
\title{Get game advanced box score information}
\usage{
cfb_game_box_advanced(game_id)
}
\arguments{
\item{game_id}{(\emph{Integer} required): Game ID filter for querying a single game
Can be found using the `cfb_game_info()` function}
}
\description{
Get game advanced box score information
}
\examples{

cfb_game_box_advanced(401012356)


}
\keyword{Game}
\keyword{Info}
