% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_info.R
\name{cfb_game_info}
\alias{cfb_game_info}
\title{Get results information from games}
\usage{
cfb_game_info(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  game_id = NULL,
  quarter_scores = FALSE
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default regular): Select Season Type: regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{home_team}{(\emph{String} optional): Home D-I Team}

\item{away_team}{(\emph{String} optional): Away D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game\cr
Can be found using the \code{\link{cfb_game_info()}} function}

\item{quarter_scores}{(\emph{Logical} default FALSE): This is a parameter to return the
list columns that give the score at each quarter: home_line_scores and away_line scores.\cr
I have defaulted the parameter to false so that you will not have to go to the trouble of dropping it.}
}
\description{
Get results information from games
}
\examples{

cfb_game_info(2018, week = 1)

cfb_game_info(2018, week = 7, conference = 'Ind')

#7 OTs LSU @ TAMU
cfb_game_info(2018, week = 13, team = "Texas A&M", quarter_scores=TRUE)

}
\keyword{Game}
\keyword{Info}
