% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_media.R
\name{cfb_game_media}
\alias{cfb_game_media}
\title{Get Game media information (TV, radio, etc)}
\usage{
cfb_game_media(
  year,
  week = NULL,
  season_type = "both",
  team = NULL,
  conference = NULL,
  media_type = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{season_type}{(\emph{String} default both): Select Season Type, regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional):  Conference Name - select a valid FBS conference
Conference Names P5: ACC,  Big 12, Big Ten, SEC, Pac-12
Conference Names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic}

\item{media_type}{(\emph{String} optional): Media type filter: tv, radio, web, ppv, or mobile}
}
\description{
Get Game media information (TV, radio, etc)
}
\keyword{Game}
\keyword{Info}
