% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_records.R
\name{cfb_game_records}
\alias{cfb_game_records}
\title{Get Team records by year}
\usage{
cfb_game_records(year, team = NULL, conference = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{team}{(\emph{String} optional): Team - Select a valid team, D1 football}

\item{conference}{(\emph{String} optional): Conference name - select a valid FBS conference\cr
Conference names P5: ACC,  Big 12, Big Ten, SEC, Pac-12\cr
Conference names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic\cr}
}
\description{
Get Team records by year
}
\examples{

cfb_game_records(2018, team = 'Notre Dame')


cfb_game_records(2013, team = "Florida State")

}
\keyword{Info}
\keyword{Team}
