% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_game_team_stats.R
\name{cfb_game_team_stats}
\alias{cfb_game_team_stats}
\title{Get Team Statistics by Game}
\usage{
cfb_game_team_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  game_id = NULL,
  rows_per_team = 1
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{season_type}{(\emph{String} default: regular): Select Season Type - regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC}

\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game
Can be found using the `cfb_game_info()` function}

\item{rows_per_team}{(\emph{Integer} default 1): Both Teams for each game on one or two row(s), Options: 1 or 2}
}
\description{
Get Team Statistics by Game
}
\examples{

cfb_game_team_stats(2018, week = 1, conference = 'B12')

cfb_game_team_stats(2019, team = 'LSU')

cfb_game_team_stats(2013, team = "Florida State")

}
\keyword{Game}
\keyword{Stats}
\keyword{Team}
