% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_metrics_ppa_players_season.R
\name{cfb_metrics_ppa_players_season}
\alias{cfb_metrics_ppa_players_season}
\title{Get player season averages for Predicted Points Added (PPA)}
\source{
\url{https://api.collegefootballdata.com/ppa/players/season}
}
\usage{
cfb_metrics_ppa_players_season(
  year = NULL,
  team = NULL,
  conference = NULL,
  position = NULL,
  athlete_id = NULL,
  threshold = NULL,
  excl_garbage_time = FALSE
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - S&P+ information by conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{position}{(\emph{string} optional): Position abbreviation of the player you are searching for.\cr
Position Group  - options include:\cr
\itemize{
\item Offense: QB, RB, FB, TE,  OL, G, OT, C, WR\cr
\item Defense: DB, CB, S, LB,  DE, DT, NT, DL\cr
\item Special Teams: K, P, LS, PK\cr
}}

\item{athlete_id}{(\emph{Integer} optional): Athlete ID filter for querying a single athlete\cr
Can be found using the \code{\link[=cfb_player_info]{cfb_player_info()}} function.}

\item{threshold}{(\emph{Integer} optional): Minimum threshold of plays.}

\item{excl_garbage_time}{(\emph{Logical} default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)}
}
\value{
A data frame with 23 variables:
\describe{
\item{\code{season}}{integer.}
\item{\code{id}}{character.}
\item{\code{name}}{character.}
\item{\code{position}}{character.}
\item{\code{team}}{character.}
\item{\code{conference}}{character.}
\item{\code{countable_plays}}{integer.}
\item{\code{avg_PPA_all}}{double.}
\item{\code{avg_PPA_pass}}{double.}
\item{\code{avg_PPA_rush}}{double.}
\item{\code{avg_PPA_first_down}}{double.}
\item{\code{avg_PPA_second_down}}{double.}
\item{\code{avg_PPA_third_down}}{double.}
\item{\code{avg_PPA_standard_downs}}{double.}
\item{\code{avg_PPA_passing_downs}}{double.}
\item{\code{total_PPA_all}}{double.}
\item{\code{total_PPA_pass}}{double.}
\item{\code{total_PPA_rush}}{double.}
\item{\code{total_PPA_first_down}}{double.}
\item{\code{total_PPA_second_down}}{double.}
\item{\code{total_PPA_third_down}}{double.}
\item{\code{total_PPA_standard_downs}}{double.}
\item{\code{total_PPA_passing_downs}}{double.}
}
}
\description{
Get player season averages for Predicted Points Added (PPA)
}
\examples{

cfb_metrics_ppa_players_season(year = 2019, team = 'TCU')

}
\keyword{Averages}
\keyword{Players}
\keyword{Points}
\keyword{Predicted}
\keyword{Season}
