% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_metrics_ppa_predicted.R
\name{cfb_metrics_ppa_predicted}
\alias{cfb_metrics_ppa_predicted}
\title{Calculate Predicted Points using Down and Distance}
\source{
\url{https://api.collegefootballdata.com/ppa/predicted}
}
\usage{
cfb_metrics_ppa_predicted(down, distance)
}
\arguments{
\item{down}{(\emph{Integer} required): Down filter}

\item{distance}{(\emph{Integer} required): Distance filter}
}
\value{
A data frame with 2 variables:
\describe{
  \item{\code{yard_line}}{integer.}
  \item{\code{predicted_points}}{character.}
}
}
\description{
Calculate Predicted Points using Down and Distance
}
\examples{

cfb_metrics_ppa_predicted(down = 1, distance = 10)

cfb_metrics_ppa_predicted(down = 3, distance = 10)

}
\keyword{Points}
\keyword{Predicted}
