% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_metrics_wp.R
\name{cfb_metrics_wp}
\alias{cfb_metrics_wp}
\title{Get win probability chart data from API}
\usage{
cfb_metrics_wp(game_id, adjust_for_spread = TRUE)
}
\arguments{
\item{game_id}{(\emph{Integer} required): Game ID filter for querying a single game\cr
Can be found using the \code{\link{cfb_game_info()}} function}

\item{adjust_for_spread}{(\emph{Logical} default TRUE): Toggles pre-game spread adjustments}
}
\description{
Get win probability chart data from API
}
\examples{

cfb_metrics_wp(game_id = 401012356)


}
\keyword{Chart}
\keyword{Data}
\keyword{Probability}
\keyword{Win}
