% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_pbp_data.R
\name{cfb_pbp_data}
\alias{cfb_pbp_data}
\title{Extract CFB (D-I) Play by Play Data - For plays}
\source{
\url{https://api.collegefootballdata.com/plays}
}
\usage{
cfb_pbp_data(
  year,
  season_type = "regular",
  week = 1,
  team = NULL,
  play_type = NULL,
  epa_wpa = FALSE
)
}
\arguments{
\item{year}{Select year, (example: 2018)}

\item{season_type}{Select Season Type (regular, postseason, both)}

\item{week}{Select week, this is optional (also numeric)}

\item{team}{Select team name (example: Texas, Texas A&M, Clemson)}

\item{play_type}{Select play type (example: see the \code{\link[cfbscrapR:cfb_play_type_df]{cfbscrapR::cfb_play_type_df}})}

\item{epa_wpa}{Logical parameter (TRUE/FALSE) to return the Expected Points Added/Win Probability Added variables}
}
\description{
Extracts raw game by game data. \cr
}
\keyword{Play-by-Play}
