% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_player_info.R
\name{cfb_player_info}
\alias{cfb_player_info}
\title{Player Information Search}
\usage{
cfb_player_info(search_term, position = NULL, team = NULL, year = NULL)
}
\arguments{
\item{search_term}{(\emph{String} required): Search term for the player you are trying to look up}

\item{position}{(\emph{string} optional): Position of the player you are searching for.\cr
Position Group  - options include:\cr
 * Offense: QB, RB, FB, TE,  OL, G, OT, C, WR\cr
 * Defense: DB, CB, S, LB,  DE, DT, NT, DL\cr
 * Special Teams: K, P, LS, PK\cr}

\item{team}{(\emph{String} optional): Team - Select a valid team, D1 football}

\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY}).
If left NULL, API default will only provide results for most recent year of final rosters: 2019}
}
\description{
A player search function with \strong{search_term} as a required input. If left NULL, API default will only provide results for most recent year of final rosters: 2019
Would not recommend using the position group filter,it feels arbitrary and pernicious,
but it is there. May fortune favor you should you choose to use it.
}
\examples{

cfb_player_info(search_term = 'James', position = 'DB', team = 'Florida State', year = 2017)

cfb_player_info(search_term = 'Lawrence', team = "Clemson")

cfb_player_info(search_term = 'Tua', position = 'QB')

cfb_player_info(search_term = 'Duggan')

}
\keyword{Recruiting}
