% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_fpi.R
\name{cfb_ratings_fpi}
\alias{cfb_ratings_fpi}
\title{Get FPI historical rating data (most recent of each year)
Adapted from sabinanalytic's fork of the cfbscrapR repo}
\source{
\url{https://github.com/sabinanalytics/cfbscrapR/blob/master/R/cfb_ratings_fpi.R}
}
\usage{
cfb_ratings_fpi(year = 2019)
}
\arguments{
\item{year}{Year}
}
\description{
Get FPI historical rating data (most recent of each year)
Adapted from sabinanalytic's fork of the cfbscrapR repo
}
\examples{

cfb_ratings_fpi(year=2018)
}
\keyword{FPI}
\keyword{Ratings}
