% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_sp.R
\name{cfb_ratings_sp}
\alias{cfb_ratings_sp}
\title{Get S&P+ historical rating data}
\usage{
cfb_ratings_sp(year = NULL, team = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{team}{(\emph{String} optional): D-I Team}
}
\description{
At least one of \strong{year} or \strong{team} must be specified for the function to run
}
\examples{

cfb_ratings_sp(year = 2019)

cfb_ratings_sp(team = 'Texas A&M')

cfb_ratings_sp(year= 2019, team = "Texas")

}
\keyword{SP+}
