% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_ratings_sp_conference.R
\name{cfb_ratings_sp_conference}
\alias{cfb_ratings_sp_conference}
\title{Get conference-level S&P+ historical rating data}
\usage{
cfb_ratings_sp_conference(year = NULL, conference = NULL)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}

\item{conference}{(\emph{String} optional): Conference abbreviation - S&P+ information by conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC}
}
\description{
Get conference-level S&P+ historical rating data
}
\examples{

cfb_ratings_sp_conference(year = 2019)

cfb_ratings_sp_conference(year = 2012, conference = 'SEC')

cfb_ratings_sp_conference(year = 2016, conference = 'ACC')

}
\keyword{SP+}
