% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_recruiting_player.R
\name{cfb_recruiting_player}
\alias{cfb_recruiting_player}
\title{CFB Recruiting
Gets CFB recruiting information for a single year with filters available for team,
recruit type, state and position.}
\source{
\url{https://api.collegefootballdata.com/recruiting/players}
}
\usage{
cfb_recruiting_player(
  year = NULL,
  team = NULL,
  recruit_type = "HighSchool",
  state = NULL,
  position = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY}) - Minimum: 2000, Maximum: 2020 currently}

\item{team}{(\emph{String} optional): D-I Team}

\item{recruit_type}{(\emph{String} optional): default API return is 'HighSchool', other options include 'JUCO'
or 'PrepSchool'  - For position group information}

\item{state}{(\emph{String} optional): Two letter State abbreviation}

\item{position}{(\emph{String} optional): Position Group  - options include:\cr
\itemize{
\item Offense: 'PRO', 'DUAL', 'RB', 'FB', 'TE',  'OT', 'OG', 'OC', 'WR'\cr
\item Defense: 'CB', 'S', 'OLB', 'ILB', 'WDE', 'SDE', 'DT'\cr
\item Special Teams: 'K', 'P'\cr
}}
}
\value{
A data frame with 14 variables:
\describe{
\item{\code{recruit_type}}{character.}
\item{\code{year}}{integer.}
\item{\code{ranking}}{integer.}
\item{\code{name}}{character.}
\item{\code{school}}{character.}
\item{\code{committed_to}}{character.}
\item{\code{position}}{character.}
\item{\code{height}}{double.}
\item{\code{weight}}{integer.}
\item{\code{stars}}{integer.}
\item{\code{rating}}{double.}
\item{\code{city}}{character.}
\item{\code{state_province}}{character.}
\item{\code{country}}{character.}
}
}
\description{
At least one of \strong{year} or \strong{team} must be specified for the function to run
}
\details{
If you would like CFB recruiting information for teams, please see the \code{\link[=cfb_recruiting_team]{cfb_recruiting_team()}}  function

If you would like to get cfb recruiting information based on position groups during a
time period for all FBS teams, please see the \code{\link[=cfb_recruiting_position]{cfb_recruiting_position()}} function.
}
\examples{

cfb_recruiting_player(2018, team = "Texas")

cfb_recruiting_player(2016, recruit_type = 'JUCO')

cfb_recruiting_player(2020, recruit_type = 'HighSchool', position = 'OT', state = 'FL')

}
\keyword{Recruiting}
