% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_game_advanced.R
\name{cfb_stats_game_advanced}
\alias{cfb_stats_game_advanced}
\title{Get Game Advanced Stats}
\usage{
cfb_stats_game_advanced(
  year,
  week = NULL,
  team = NULL,
  opponent = NULL,
  excl_garbage_time = FALSE,
  season_type = "both"
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}

\item{week}{(\emph{Integer} optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)}

\item{team}{(\emph{String} optional): D-I Team}

\item{opponent}{(\emph{String} optional): Opponent D-I Team}

\item{excl_garbage_time}{(\emph{Logical} default FALSE): Select whether to exclude Garbage Time (TRUE/FALSE)}

\item{season_type}{(\emph{String} default both): Select Season Type: regular, postseason, or both.}
}
\description{
Get Game Advanced Stats
}
\examples{

cfb_stats_game_advanced(year = 2018, week = 12, team = 'Texas A&M')

cfb_stats_game_advanced(2019, team = 'LSU')

cfb_stats_game_advanced(2013, team = "Florida State")

}
\keyword{Advanced}
\keyword{Game}
\keyword{Stats}
