% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_stats_season_team.R
\name{cfb_stats_season_team}
\alias{cfb_stats_season_team}
\title{Get Season Statistics by Team}
\usage{
cfb_stats_season_team(
  year,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  start_week = NULL,
  end_week = NULL
)
}
\arguments{
\item{year}{(\emph{Integer} required): Year, 4 digit format (\emph{YYYY})}

\item{season_type}{(\emph{String} default: regular): Select Season Type - regular, postseason, or both}

\item{team}{(\emph{String} optional): D-I Team}

\item{conference}{(\emph{String} optional): Conference abbreviation - Select a valid FBS conference\cr
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC\cr
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC\cr}

\item{start_week}{(\emph{Integer} optional): Starting Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}

\item{end_week}{(\emph{Integer} optional): Ending Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier}
}
\description{
Get Season Statistics by Team
}
\examples{

cfb_stats_season_team(year=2018, conference = 'B12', start_week = 1,end_week = 8)

cfb_stats_season_team(2019, team = 'LSU')

cfb_stats_season_team(2013, team = "Florida State")

}
\keyword{Season}
\keyword{Stats}
\keyword{Team}
