% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epa.R
\name{create_epa}
\alias{create_epa}
\title{Create EPA
Adds Expected Points calculations to Play-by-Play data.frame}
\usage{
create_epa(
  clean_pbp_dat,
  ep_model = cfbscrapR:::ep_model,
  fg_model = cfbscrapR:::fg_model
)
}
\arguments{
\item{clean_pbp_dat}{(\emph{data.frame} required): Clean PBP as input from \code{\link[cfbscrapR:cfb_pbp_data]{cfbscrapR::cfb_pbp_data()}})}

\item{ep_model}{(\emph{model} default cfbscrapR:ep_model): Expected Points (EP) Model}

\item{fg_model}{(\emph{model} default cfbscrapR:fg_model): Field Goal (FG) Model}
}
\description{
Create EPA
Adds Expected Points calculations to Play-by-Play data.frame
}
\keyword{internal}
