% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epa.R
\name{prep_epa_df_after}
\alias{prep_epa_df_after}
\title{Creates the post-play inputs for the Expected Points model to predict on for each game}
\usage{
prep_epa_df_after(dat)
}
\arguments{
\item{dat}{(\emph{Data.Frame} required) Clean Play-by-Play DataFrame pulled from `cfb_pbp_dat()`}
}
\description{
Creates the post-play inputs for the Expected Points model to predict on for each game
}
\keyword{internal}
